fix: temp dir name chnage to fix casing
