+++ src/components/templates/appBarTemp/AppBar.tsx
@@ -4,11 +4,11 @@ import { Box } from "@mui/system";
 import styles from "./AppBar.module.css";
 
 export const CustomAppBar = () => {
-    return (
-        <Box>
-            <AppBar position="fixed" className={styles.appbar}>
-                <CustomToolbar />
-            </AppBar>
-        </Box>
-    );
+  return (
+    <Box>
+      <AppBar position="fixed" className={styles.appbar}>
+        <CustomToolbar />
+      </AppBar>
+    </Box>
+  );
 };
