--- src/components/templates/hero/index.ts
@@ -1,1 +1,0 @@
-export * from "./Hero";
