+++ src/components/templates/heroTemp/index.ts
